Minor Change for ui class to formatter class as it no longer handles ui. Refactoring of rest of the classes who use methods in the Formatter class
