+++ src/main/java/oreo/Main.java
@@ -6,7 +6,9 @@ import javafx.application.Application;
 import javafx.fxml.FXMLLoader;
 import javafx.scene.Scene;
 import javafx.scene.layout.AnchorPane;
+import javafx.scene.text.Font;
 import javafx.stage.Stage;
+import oreo.ui.MainWindow;
 
 /**
      * A GUI for Duke using FXML.
