+++ src/main/java/oreo/Oreo.java
@@ -4,16 +4,14 @@ import oreo.command.EditCommand;
 import oreo.exception.IllegalCommandException;
 import oreo.exception.IllegalDateTimeException;
 import oreo.command.Command;
-import oreo.parser.Parser;
 import oreo.storage.Storage;
 import oreo.task.TaskList;
 import oreo.task.Task;
-import oreo.ui.Ui;
+import oreo.ui.MainWindow;
 
 import java.io.FileNotFoundException;
 import java.io.IOException;
 import java.util.ArrayList;
-import java.util.InputMismatchException;
 
 
 /**
