+++ src/main/java/oreo/command/ByeCommand.java
@@ -2,7 +2,6 @@ package oreo.command;
 
 import oreo.task.Task;
 import oreo.task.TaskList;
-import oreo.ui.Ui;
 
 public class ByeCommand extends Command {
