+++ src/main/java/oreo/command/Command.java
@@ -4,7 +4,6 @@ import oreo.exception.IllegalCommandException;
 import oreo.exception.IllegalDateTimeException;
 import oreo.task.Task;
 import oreo.task.TaskList;
-import oreo.ui.Ui;
 
 /**
  * Implements abstract command class, parent class of specific commands
