+++ src/main/java/oreo/ui/DialogBox.java
@@ -1,4 +1,4 @@
-package oreo;
+package oreo.ui;
 
 import java.io.IOException;
 import java.util.Collections;
@@ -12,13 +12,9 @@ import javafx.scene.Node;
 import javafx.scene.control.Label;
 import javafx.scene.image.Image;
 import javafx.scene.image.ImageView;
-import javafx.scene.layout.Background;
-import javafx.scene.layout.BackgroundFill;
 import javafx.scene.layout.HBox;
 import javafx.scene.paint.Color;
-import oreo.MainWindow;
-
-import static com.sun.javafx.fxml.expression.Expression.set;
+import javafx.scene.text.Font;
 
 /**
  * An example of a custom control using FXML.
@@ -47,6 +43,8 @@ public class DialogBox extends HBox {
         dialog.setText(text);
         displayPicture.setImage(img);
         speechBubble.heightProperty();
+//        Font dmSansFont = Font.loadFont(getClass().getResourceAsStream("/fonts/DMSans-Medium.ttf"), 11);
+//        dialog.setFont(dmSansFont);
     }
 
     /**
